@@ -14,7 +14,7 @@ description: SakhaTyla app
 version: 2.0.4+204
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.5.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,9 +26,9 @@ dependencies:
   http: ^0.12.0+4
   html: ^0.14.0+3
   get_it: ^3.1.0
-  flutter_bloc: ^3.2.0
-  equatable: ^1.1.1
-  url_launcher: ^5.4.10
+  flutter_bloc: ^6.0.0
+  equatable: ^1.2.0
+  url_launcher: ^5.5.2
 
 dev_dependencies:
   flutter_test:
